Refactor processing chain test Fix columns() in test
